fixed metric test (oops, forgot to run this before checking in)

+++ dao/src/test/java/com/inthinc/pro/dao/util/MetricConversionUtilTest.java
@@ -60,7 +60,7 @@ public class MetricConversionUtilTest
     {
         Float expectedKilometers = 27.36F;
         Double actualMiles = 17D;
-        Float kilometers = MeasurementConversionUtil.fromMilesToKilometers(actualMiles);        
+        Float kilometers = MeasurementConversionUtil.fromMilesToKilometers(actualMiles).floatValue();        
         Assert.assertEquals(expectedKilometers.toString(), kilometers.toString());
     }
